requirements.txt
attrs==21.4.0
cmd2==2.4.0
pyperclip==1.8.2
wcwidth==0.2.5
discord.py==1.7.3
pymediawiki==0.7.1
